requirement_linux.txt
alabaster==0.7.10
amqp==1.4.9
anyjson==0.3.3
astroid==1.5.2
Babel==2.5.3
backcall==0.1.0
baidu-aip==2.2.4.0
beautifulsoup4==4.5.1
billiard==3.3.0.23
boto==2.42.0
cached-property==1.4.0
celery==3.1.26.post2
chardet==3.0.4
click==6.7
click-log==0.1.8
cookies==2.2.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.2
ddt==1.1.0
decorator==4.4.0
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.11.11
django-appconf==1.0.2
django-autocomplete-light==3.1.8
django-celery==3.3.1
django-celery-results==1.0.0
django-choices==1.4.3
django-compressor==2.1.1
django-contrib-comments==1.8.0
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.8
django-elasticsearch-debug-toolbar==1.2.0
django-extensions==1.7.8
django-filter==1.0.4
django-formtools==2.1
django-fsm==2.6.0
django-gm2m==0.6.1
django-guardian==1.4.8
django-haystack==2.8.1
django-imagekit==4.0.2
django-import-export==1.0.0
django-ipware==2.1.0
django-js-asset==1.1.0
django-libsass==0.7
django-model-utils==3.1.2
django-modelcluster==3.1
django-mptt==0.9.0
django-parler==1.8
django-pdb==0.6.2
django-ratelimit-backend==1.0
django-rest-swagger==2.1.2
django-reversion==2.0.13
django-ses==0.8.1
django-simple-history==1.9.0
django-simpleui==3.9.1
django-solo==1.1.2
django-sortedm2m==1.4.0
django-statici18n==1.7.0
django-stdimage==2.4.1
django-storages==1.5.2
django-taggit==0.22.1
django-taggit-autosuggest==0.3.0
django-taggit-serializer==0.1.5
django-treebeard==4.2.1
django-waffle==0.11.1
django-webpack-loader==0.5.0
djangorestframework==3.6.3
djangorestframework-csv==1.4.1
djangorestframework-jwt==1.8.0
djangorestframework-xml==1.3.0
docker==2.7.0
docker-compose==1.19.0
docker-pycreds==0.2.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
draftjs-exporter==2.0.0
drf-extensions==0.3.1
drf-haystack==1.6.1
dry-rest-permissions==0.1.6
edx-auth-backends==1.1.3
edx-ccx-keys==0.2.0
edx-django-release-util==0.3.1
edx-django-sites-extensions==2.3.0
edx-drf-extensions==1.2.3
edx-i18n-tools==0.4.4
edx-opaque-keys==0.4.4
elasticsearch==1.9.0
et-xmlfile==1.0.1
factory-boy==2.8.1
Faker==0.8.12
freezegun==0.3.7
future==0.16.0
gevent==1.1.2
greenlet==0.4.13
gunicorn==19.6.0
html2text==2016.9.19
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
imagesize==1.0.0
importlib-metadata==1.5.0
ipip-ipdb==1.3.1
ipython==7.6.0
ipython-genutils==0.2.0
isort==4.2.5
itypes==1.1.0
jdcal==1.3
jedi==0.14.0
Jinja2==2.10
joblib==0.14.1
jsonfield==1.0.3
jsonschema==2.6.0
kombu==3.0.37
lazy-object-proxy==1.3.1
libsass==0.14.2
lxml==3.6.1
Markdown==2.6.6
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
mysqlclient==1.3.9
newrelic==2.86.3.70
nltk==3.5
oauthlib==2.0.7
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.2
parso==0.5.0
path.py==11.0
pbr==3.1.1
pep8==1.7.0
pexpect==4.7.0
pickleshare==0.7.5
pilkit==2.0
Pillow==3.4.2
polib==1.1.0
progressbar2==3.36.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.1
py4j==0.10.8.1
pyasn1==0.4.2
pycountry==1.20
pycparser==2.18
pycryptodomex==3.9.7
Pygments==2.2.0
pyjwkest==1.4.0
PyJWT==1.6.1
pylint==1.7.1
pylint-celery==0.3
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pymongo==3.6.1
PyMySQL==0.9.3
pytest==3.0.6
pytest-django==3.2.1
pytest-django-ordering==1.0.1
pytest-responses==0.3.0
python-dateutil==2.5.3
python-memcached==1.58
python-utils==2.3.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.12
rcssmin==1.0.6
redis==3.4.1
regex==2020.4.4
requests==2.9.1
requests-oauthlib==0.8.0
responses==0.7.0
rjsmin==1.0.12
selenium==3.4.0
simplejson==3.13.2
six==1.11.0
slumber==0.7.1
snowballstemmer==1.2.1
social-auth-app-django==1.2.0
social-auth-core==1.7.0
Sphinx==1.5.3
sqlparse==0.2.4
stevedore==1.28.0
supervisor==4.1.0
tablib==0.12.1
testfixtures==4.13.1
text-unidecode==1.2
texttable==0.9.1
tqdm==4.45.0
traitlets==4.3.2
transifex-client==0.12.2
unicodecsv==0.14.1
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
vine==1.3.0
virtualenv==15.1.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.47.0
Willow==1.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
zipp==1.2.0
